Build 119: Ambient Piano with SC55 Stereo piano
- Ambient Piano: SC55 Stereo piano (Roland SC55 soundfont) replaces
  GeneralUser GS Grand Piano as the default Lead 1 instrument — more
  dynamic and natural-sounding. Boost via AVAudioMixerNode to bring
  sample level up to mix.
- Ambient Lead 1 pool simplified to: Stereo Piano, Flute, Ocarina,
  Whistle, Brightness, Calliope Lead, Harp.
- Chill Blues bass and other musical improvements.
- LFO boost node fix: instrument-level gain now routed through
  boostBaseMultiplier on the mixer node rather than AVAudioUnitSampler
  volume (which is clamped to [0,1] by Apple).
- release-dmg.sh updated to version 1.5.

+++ Sources/Zudio/Assets/SongLogExporter.swift
@@ -187,12 +187,12 @@ struct SongLogExporter {
         return parts.joined(separator: " ")
     }
 
-    private static func generationProgram(forTrack i: Int, song: SongState) -> UInt8 {
+    private static func generationProgram(forTrack i: Int, song: SongState) -> Int {
         if song.style == .chill {
-            if i == kTrackLead1 { return song.chillLeadInstrument.gmProgram }
-            if i == kTrackLead2 { return song.chillLead2Instrument.gmProgram }
+            if i == kTrackLead1 { return Int(song.chillLeadInstrument.gmProgram) }
+            if i == kTrackLead2 { return Int(song.chillLead2Instrument.gmProgram) }
         }
-        if i == kTrackLeadSynth { return kDefaultGMPrograms[kTrackLeadSynth] ?? 90 }
+        if i == kTrackLeadSynth { return Int(kDefaultGMPrograms[kTrackLeadSynth] ?? 90) }
         let pool = AppState.instrumentPoolPrograms(trackIndex: i, style: song.style)
         return pool.first ?? 255
     }
